[svn-r1010] Changes since 19981217 ----------------------
./src/H5.c
./src/H5private.h
        Renamed `library_initialize_g' to `H5_libinit_g' to make it
        conform to our naming scheme.

./src/H5I.c
        Fixed a bug in H5I_dec_ref() that caused the return value to
        always be zero instead of the new reference count.

./src/H5.c
./src/H5Odtype.c
./src/H5T.c
./src/H5Tconv.c
./src/H5Tpkg.h
./src/H5Tprivate.h
./src/H5Tpublic.h
        Added support for enumeration data types.

./src/H5RA.c
        Renamed H5T_insert() to H5T_struct_insert() and added
        H5T_enum_insert().

./src/H5RA.c
./src/H5Shyper.c
        Added casts to size_t for the third argument of memcpy() to
        shut up a warning message from insure++.

./src/H5T.c
        Changed "can't" to "unable to" in some error messages to be
        more consistent.

./src/H5detect.c
        If fork() or waitpid() are unavailable then we assume no
        alignment constraints. Hopefully this is the case only on NT
        and Intel CPU's don't have alignment constraints.

./src/H5public.h
        Include <limits.h> because the H5T_NATIVE_CHAR macro needs the
        definition for CHAR_MIN.

+++ src/H5Shyper.c
@@ -503,7 +503,7 @@ H5S_hyper_block_read (H5S_hyper_node_t *node, H5S_hyper_fhyper_info_t *fhyper_in
     */
     HDmemcpy(fhyper_info->dst,
 	     node->cinfo.pos,
-	     region_size*fhyper_info->elmt_size);
+	     (size_t)(region_size*fhyper_info->elmt_size));
 
     /*
      * Decrement the number of elements left in block to read & move the
@@ -559,7 +559,7 @@ H5S_hyper_block_write (H5S_hyper_node_t *node,
     */
     HDmemcpy(node->cinfo.pos,
 	     fhyper_info->src,
-	     region_size*fhyper_info->elmt_size);
+	     (size_t)(region_size*fhyper_info->elmt_size));
 
     /*
      * Decrement the number of elements left in block to read & move the
